% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.character.bm_glyph.R
\name{as.character.bm_glyph}
\alias{as.character.bm_glyph}
\title{Coerce bitmap glyph objects to character vector}
\usage{
\method{as.character}{bm_glyph}(x, ..., labels = c(" ", "█", "▒"))
}
\arguments{
\item{x}{A \code{bm_glyph()} object}

\item{...}{Further arguments passed to or from other methods.}

\item{labels}{Character vector of the labels to use for each integer value i.e.
The first character for integer \code{0L},
the second character for integer \code{1L}, and so on.}
}
\value{
A character vector of the string representation.
}
\description{
\code{as.character.bm_glyph()} converts bitmap glyph objects
to a character vector.
}
\examples{
  plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
  plus_sign[5L, 3:7] <- 1L
  plus_sign[3:7, 5L] <- 1L
  plus_sign <- bm_glyph(plus_sign)
  as.character(plus_sign, labels = c(" ", "#"))

  space_glyph <- bm_glyph(matrix(0L, nrow = 8L, ncol = 8L))
  as.character(space_glyph, labels = ".")
}
\seealso{
[bm_glyph(), \code{\link[=print.bm_glyph]{print.bm_glyph()}}
}
