% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.bm_bitmap.R
\name{as.matrix.bm_bitmap}
\alias{as.matrix.bm_bitmap}
\title{Coerce bitmap objects to matrix}
\usage{
\method{as.matrix}{bm_bitmap}(x, ...)
}
\arguments{
\item{x}{A \code{bm_bitmap()} object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An integer matrix
}
\description{
\code{as.matrix.bm_bitmap()} coerces \code{bm_bitmap()} objects to an integer matrix.
}
\examples{
 space_matrix <- matrix(0L, ncol = 8L, nrow = 8L)
 space_glyph <- bm_bitmap(space_matrix)
 print(space_glyph, px = ".")
 print(as.matrix(space_glyph))
}
