% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.bm_glyph.R
\name{as.matrix.bm_glyph}
\alias{as.matrix.bm_glyph}
\title{Coerce bitmap glyph objects to matrix}
\usage{
\method{as.matrix}{bm_glyph}(x, ...)
}
\arguments{
\item{x}{A \code{bm_glyph()} object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An integer matrix
}
\description{
\code{as.matrix.bm_glyph()} coerces \code{bm_glyph()} objects to an integer matrix.
}
\examples{
 space_matrix <- matrix(0L, ncol = 8L, nrow = 8L)
 space_glyph <- bm_glyph(space_matrix)
 print(space_glyph, labels = ".")
 print(as.matrix(space_glyph))
}
