% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.bm_matrix.R
\name{as.matrix.bm_matrix}
\alias{as.matrix.bm_matrix}
\alias{as.matrix.bm_bitmap}
\alias{as.matrix.bm_pixmap}
\title{Cast bitmap/pixmap objects to a (normal) matrix}
\usage{
\method{as.matrix}{bm_matrix}(x, ...)

\method{as.matrix}{bm_matrix}(x, ...)
}
\arguments{
\item{x}{Either a \code{\link[=bm_bitmap]{bm_bitmap()}} or \code{\link[=bm_pixmap]{bm_pixmap()}} object.}

\item{...}{Currently ignored.}
}
\value{
Either an integer matrix if \code{x} is a \code{\link[=bm_bitmap]{bm_bitmap()}} object
or a character matrix if \code{x} is a \code{\link[=bm_pixmap]{bm_pixmap()}} object.
}
\description{
\code{as.matrix.bm_matrix()} casts \code{\link[=bm_bitmap]{bm_bitmap()}} objects to an (normal) integer matrix.
and \code{\link[=bm_pixmap]{bm_pixmap()}} objects to a (normal) character matrix (of color strings).
}
\examples{
 space_matrix <- matrix(0L, ncol = 8L, nrow = 8L)
 space_glyph <- bm_bitmap(space_matrix)
 print(space_glyph, px = ".")
 print(as.matrix(space_glyph))
}
