% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.raster.bm_bitmap.R
\name{as.raster.bm_bitmap}
\alias{as.raster.bm_bitmap}
\title{Coerce bitmap glyph object to raster object}
\usage{
\method{as.raster}{bm_bitmap}(x, ..., col = c("grey80", "black", "grey40"))
}
\arguments{
\item{x}{A \code{bm_bitmap()} object}

\item{...}{Further arguments passed to or from other methods.}

\item{col}{Character vector of R color specifications}
}
\description{
\code{as.raster.bm_bitmap()} converts a bitmap glyph object to a raster object.
}
\examples{
  plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
  plus_sign[5L, 3:7] <- 1L
  plus_sign[3:7, 5L] <- 1L
  plus_sign <- bm_bitmap(plus_sign)
  grDevices::as.raster(plus_sign)
}
