% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_font.R
\name{as_bm_font}
\alias{as_bm_font}
\alias{as_bm_font.list}
\alias{as_bm_font.default}
\title{Coerce to bitmap font objects}
\usage{
as_bm_font(x, ...)

\method{as_bm_font}{list}(x, ...)

\method{as_bm_font}{default}(x, ...)
}
\arguments{
\item{x}{An object that can reasonably be coerced to a \code{bm_font()} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{bm_font()} object.
}
\description{
\code{as_bm_font()} turns an existing object into a \code{bm_font()} object.
}
\examples{
  plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
  plus_sign[5L, 3:7] <- 1L
  plus_sign[3:7, 5L] <- 1L
  plus_sign_glyph <- bm_glyph(plus_sign)
  plus_sign_code_point <- code_point_from_name("PLUS SIGN") # "U+002B"

  space_glyph <- bm_glyph(matrix(0L, nrow = 9L, ncol = 9L))
  space_code_point <- code_point_from_name("SPACE") # "U+0020"

  l <- list()
  l[[plus_sign_code_point]] <- plus_sign_glyph
  l[[space_code_point]] <- space_glyph
  font <- as_bm_font(l)
  is_bm_font(font)

}
\seealso{
\code{\link[=bm_font]{bm_font()}}
}
