% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_glyph.R
\name{as_bm_glyph}
\alias{as_bm_glyph}
\alias{as_bm_glyph.matrix}
\alias{as_bm_glyph.default}
\title{Coerce to bitmap glyph objects}
\usage{
as_bm_glyph(x, ...)

\method{as_bm_glyph}{matrix}(x, ...)

\method{as_bm_glyph}{default}(x, ...)
}
\arguments{
\item{x}{An object that can reasonably be coerced to a \code{bm_glyph()} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{bm_glyph()} object.
}
\description{
\code{as_bm_glyph()} turns an existing object into a \code{bm_glyph()} object.
}
\examples{
 space_matrix <- matrix(0, nrow = 16, ncol = 16)
 space_glyph <- as_bm_glyph(space_matrix)
 is_bm_glyph(space_glyph)
}
\seealso{
\code{\link[=bm_glyph]{bm_glyph()}}
}
