% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_list.R
\name{as_bm_list}
\alias{as_bm_list}
\alias{as_bm_list.default}
\alias{as_bm_list.list}
\title{Coerce to bitmap glyph list objects}
\usage{
as_bm_list(x, ...)

\method{as_bm_list}{default}(x, ...)

\method{as_bm_list}{list}(x, ...)
}
\arguments{
\item{x}{An object that can reasonably be coerced to a \code{\link[=bm_list]{bm_list()}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{\link[=bm_list]{bm_list()}} object.
}
\description{
\code{as_bm_list()} turns an existing object into a \code{\link[=bm_list]{bm_list()}} object.
}
\examples{
  plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
  plus_sign[5L, 3:7] <- 1L
  plus_sign[3:7, 5L] <- 1L
  plus_sign_glyph <- bm_bitmap(plus_sign)

  space_glyph <- bm_bitmap(matrix(0L, nrow = 9L, ncol = 9L))

  l <- list()
  l[[str2ucp("+")]] <- plus_sign_glyph
  l[[str2ucp(" ")]] <- space_glyph
  bl <- as_bm_list(l)
  is_bm_list(bl)

}
\seealso{
\code{\link[=bm_list]{bm_list()}}
}
