% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{cbind.bm_bitmap}
\alias{cbind.bm_bitmap}
\alias{rbind.bm_bitmap}
\title{Combine bitmaps by rows or columns}
\usage{
\method{cbind}{bm_bitmap}(...)

\method{rbind}{bm_bitmap}(...)
}
\arguments{
\item{...}{\code{\link[=bm_bitmap]{bm_bitmap()}} objects.}
}
\value{
A \code{\link[=bm_bitmap]{bm_bitmap()}} object.
}
\description{
\code{cbind.bm_bitmap()} and \code{rbind.bm_bitmap()} combine
by columns or rows respectively.
}
\examples{
  font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
  font <- read_hex(font_file)
  capital_b <- font[[str2ucp("B")]]
  capital_m <- font[[str2ucp("M")]]
  cbm <- cbind(capital_b, capital_m)
  print(cbm, labels = c(".", "#"))
  rbm <- rbind(capital_b, capital_m)
  print(rbm, labels = c(".", "#"))
}
