% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_bytepad.R
\name{bm_bytepad}
\alias{bm_bytepad}
\title{Pad bitmap widths to the nearest byte}
\usage{
bm_bytepad(x, ...)
}
\arguments{
\item{x}{Either a \code{\link[=bm_bitmap]{bm_bitmap()}}, \code{\link[=bm_font]{bm_font()}}, \code{\link[=bm_list]{bm_list()}}, \link[magick:editing]{"magick-image"}, "nativeRaster", \code{\link[=bm_pixmap]{bm_pixmap()}}, or \link[grDevices:as.raster]{"raster"}  object.}

\item{...}{Passed to \code{\link[=bm_extend]{bm_extend()}} along with \code{hjust = "left"} and \code{width_multiples_of = 8L}.}
}
\value{
Depending on \code{x} either a \code{\link[=bm_bitmap]{bm_bitmap()}}, \code{\link[=bm_font]{bm_font()}}, \code{\link[=bm_list]{bm_list()}}, \link[magick:editing]{magick-image}, "nativeRaster", \code{\link[=bm_pixmap]{bm_pixmap()}}, or \link[grDevices:as.raster]{raster}  object.
}
\description{
\code{bm_bytepad()} pads the width of bitmaps to the nearest multiple of 8
by padding with pixels on the right.
}
\details{
This is required by the BDF font format, which specifies that each row of
bitmap data must be padded with zero bits on the right to the nearest byte.
}
\examples{
font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
font <- read_hex(font_file)
capital_r <- font[[str2ucp("R")]]
ncol(capital_r) # already 8 pixels wide
capital_r_7 <- bm_trim(capital_r, right = 1L)
ncol(capital_r_7) # trimmed to 7 pixels wide
capital_r_8 <- bm_bytepad(capital_r_7)
ncol(capital_r_8) # padded back to 8 pixels wide
all.equal(capital_r, capital_r_8)
}
\seealso{
\code{\link[=bm_extend]{bm_extend()}}
}
