% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_gray.R
\name{bm_gray}
\alias{bm_gray}
\alias{bm_grey}
\alias{bm_gray.bm_bitmap}
\alias{bm_gray.bm_list}
\alias{bm_gray.bm_pixmap}
\alias{bm_gray.magick-image}
\alias{bm_gray.nativeRaster}
\alias{bm_gray.raster}
\title{Gray a bitmap}
\usage{
bm_gray(x)

bm_grey(x)

\method{bm_gray}{bm_bitmap}(x)

\method{bm_gray}{bm_list}(x)

\method{bm_gray}{bm_pixmap}(x)

\method{bm_gray}{`magick-image`}(x)

\method{bm_gray}{nativeRaster}(x)

\method{bm_gray}{raster}(x)
}
\arguments{
\item{x}{Either a \code{\link[=bm_bitmap]{bm_bitmap()}}, \code{\link[=bm_font]{bm_font()}}, \code{\link[=bm_list]{bm_list()}}, \link[magick:editing]{"magick-image"}, "nativeRaster", \code{\link[=bm_pixmap]{bm_pixmap()}}, or \link[grDevices:as.raster]{"raster"}  object.}
}
\value{
Depending on \code{x} either a \code{\link[=bm_bitmap]{bm_bitmap()}}, \code{\link[=bm_font]{bm_font()}}, \code{\link[=bm_list]{bm_list()}}, \link[magick:editing]{magick-image}, "nativeRaster", \code{\link[=bm_pixmap]{bm_pixmap()}}, or \link[grDevices:as.raster]{raster}  object.
}
\description{
\code{bm_gray()} grays a bitmap. \code{bm_grey()} is offered as an alias.
}
\examples{
corn <- farming_crops_16x16()$corn$portrait
corn_gray <- bm_gray(corn)
if (cli::is_utf8_output() && cli::num_ansi_colors() >= 256L) {
  print(corn_gray, compress = "v")
}
}
