% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_list.R
\name{bm_lapply}
\alias{bm_lapply}
\title{Modify bitmap lists}
\usage{
bm_lapply(X, FUN, ...)
}
\arguments{
\item{X}{A bitmap glyph list object such as \code{\link[=bm_list]{bm_list()}} or \code{\link[=bm_font]{bm_font()}}.}

\item{FUN}{A function that takes a \code{\link[=bm_bitmap]{bm_bitmap()}} object as its first argument
and returns a \code{\link[=bm_bitmap]{bm_bitmap()}} object.}

\item{...}{Additional arguments to pass to \code{FUN}.}
}
\value{
A modified bitmap glyph list.
}
\description{
\code{bm_lapply()} applies a function over a bitmap glyph list
and returns a modified bitmap glyph list.
}
\details{
\code{bm_lapply()} is a wrapper around \code{base::lapply()} that
preserves the classes and metadata of the original bitmap glyph list.
}
\seealso{
\code{\link[base:lapply]{base::lapply()}}, \code{\link[=bm_list]{bm_list()}}, \code{\link[=bm_font]{bm_font()}}, \code{\link[=bm_bitmap]{bm_bitmap()}}
}
