% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_options.R
\name{bm_options}
\alias{bm_options}
\title{Get bittermelon options}
\usage{
bm_options(..., default = FALSE)
}
\arguments{
\item{...}{\code{bittermelon} package options using \code{name = value}.
The return list will use any of these instead of the current/default values.}

\item{default}{If \code{TRUE} return the default values instead of current values.}
}
\value{
A list of option values.
Note this function \strong{does not} set option values itself but
this list can be passed to \code{\link[=options]{options()}}, \code{\link[withr:with_options]{withr::local_options()}}, or \code{\link[withr:with_options]{withr::with_options()}}.
}
\description{
\code{bm_options()} returns the \code{bittermelon} packages global options.
}
\examples{
  bm_options()

  bm_options(default = TRUE)

  bm_options(bittermelon.compress = "vertical")
}
\seealso{
\link{bittermelon} for a high-level description of relevant global options.
}
