% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_replace.R
\name{bm_replace}
\alias{bm_replace}
\alias{bm_replace.bm_bitmap}
\alias{bm_replace.bm_list}
\alias{bm_replace.bm_pixmap}
\alias{bm_replace.magick-image}
\alias{bm_replace.nativeRaster}
\alias{bm_replace.raster}
\title{Replace a color in a bitmap with another color}
\usage{
bm_replace(x, value, old)

\method{bm_replace}{bm_bitmap}(x, value = 0L, old = x[1L, 1L])

\method{bm_replace}{bm_list}(x, value = 0L, ...)

\method{bm_replace}{bm_pixmap}(x, value = col2hex("transparent"), old = x[1L, 1L])

\method{bm_replace}{`magick-image`}(x, value = "transparent", old = as.raster(x)[1L, 1L])

\method{bm_replace}{nativeRaster}(x, value = col2int("transparent"), old = x[1L, 1L])

\method{bm_replace}{raster}(x, value = "transparent", old = x[1L, 1L])
}
\arguments{
\item{x}{A bitmap object.}

\item{value}{New bitmap color value.}

\item{old}{Old bitmap color value to replace.}

\item{...}{Passed to \code{bm_replace.bm_bitmap()}.}
}
\value{
A bitmap object.
}
\description{
\code{bm_replace()} replaces a bitmap color with another color.
In particular default arguments will try to replace the background color.
}
\examples{
corn <- farming_crops_16x16()$corn$portrait
bm_replace(corn, "cyan")
bm_replace(corn, "magenta")
}
