% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_point.R
\name{code_point}
\alias{code_point}
\alias{code_point_from_char}
\alias{code_point_from_name}
\title{Get Unicode code points}
\usage{
code_point(x)

code_point_from_char(x)

code_point_from_name(x)
}
\arguments{
\item{x}{R objects coercible to the respective Unicode character data types.
See \code{\link[Unicode:u_char_basics]{Unicode::as.u_char()}} for \code{code_point()},
\code{\link[base:utf8Conversion]{base::utf8ToInt()}} for \code{code_point_from_char()}, and
\code{\link[Unicode:u_char_names]{Unicode::u_char_from_name()}} for \code{code_point_from_name()}.}
}
\value{
A character vector of Unicode code points.
}
\description{
\code{code_point()}, \code{code_point_from_char()}, and \code{code_point_from_name()} return Unicode code points.
}
\details{
\code{code_point()} is a wrapper for \code{as.character(Unicode::as.u_char())}.
\code{code_point_from_char()} is a wrapper for \code{as.character(Unicode::as.u_char(utf8ToInt()))}.
\code{code_point_from_name()} is a wrapper for \code{as.character(Unicode::u_char_from_name())}.
Note the names of \code{bm_font()} objects must be character vectors as returned
by these functions and not \code{u_char} objects.
}
\examples{
  # These are all different ways to get the 'R' code point
  code_point("52")
  code_point(as.hexmode("52"))
  code_point(82) # 82 == as.hexmode("52")
  code_point("U+0052")
  code_point("0x0052")
  code_point_from_char("R")
  code_point_from_name("LATIN CAPITAL LETTER R")
}
