% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-color.R
\name{col2hex}
\alias{col2hex}
\title{Colors to standardized hex strings}
\usage{
col2hex(x)
}
\arguments{
\item{x}{Color value as supported by \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}.}
}
\value{
A standardized RGBA hex string (as returned by \code{\link[grDevices:rgb]{grDevices::rgb()}}).
}
\description{
\code{col2hex()} standardizes R color strings
into a unique RGBA hex string.
All fully transparent colors get standardized
to \code{"#FFFFFF00"}.
}
\examples{
col2hex("red")
col2hex("green")
col2hex("blue")
col2hex("transparent")
col2hex(NA_character_)
col2hex("#00000000")
}
