% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-color.R
\name{col2int}
\alias{col2int}
\alias{int2col}
\title{Color to (native) integer conversions}
\usage{
col2int(x)

int2col(x)
}
\arguments{
\item{x}{Color value to convert.}
}
\value{
\code{col2int()} returns an integer.  \code{int2col()} returns a (hex) color string.
}
\description{
\code{col2int()} converts color strings to (native) color integers.
\code{int2col()} converts (native) color integers to color strings.
}
\details{
\itemize{
\item Colors are also standardized by \code{\link[=col2hex]{col2hex()}}.
\item Requires the \link[farver:farver-package]{farver} package.
}
}
\examples{
if (requireNamespace("farver", quietly = TRUE)) {
  int2col(col2int("red"))
}
}
