% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_crop.R
\name{food_crop_sprites}
\alias{food_crop_sprites}
\title{Sprites for twenty food crops}
\usage{
food_crop_sprites()
}
\value{
A named list of lists of six \code{\link[=bm_pixmap]{bm_pixmap()}} objects (one through five stages of growth plus a portrait for each crop).  The named list has the following twenty crop names:
\itemize{
\item "avocado"
\item "cassava"
\item "coffee"
\item "corn"
\item "cucumber"
\item "eggplant"
\item "grapes"
\item "lemon"
\item "melon"
\item "orange"
\item "pineapple"
\item "potato"
\item "rice"
\item "rose"
\item "strawberry"
\item "sunflower"
\item "tomato"
\item "tulip"
\item "turnip"
\item "wheat"
}
}
\description{
\code{food_crop_sprites()} returns a named list of lists
of twenty food crops in six various stages as \code{\link[=bm_pixmap]{bm_pixmap()}} objects.
}
\examples{
crops <- food_crop_sprites()
names(crops)
if (cli::is_utf8_output() && cli::num_ansi_colors() >= 256L) {
  print(crops$corn[[6L]], compress = "v")
}
if (cli::is_utf8_output() && cli::num_ansi_colors() >= 256L) {
  print(crops$orange[[5L]], compress = "v")
}
}
