% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bm_matrix.R
\name{is_bm_bitmap}
\alias{is_bm_bitmap}
\title{Test if the object is a bitmap object}
\usage{
is_bm_bitmap(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{is_bm_bitmap()} returns \code{TRUE} for \code{bm_bitmap} objects (or subclasses)
and \code{FALSE} for all other objects.
}
\examples{
space_matrix <- matrix(0L, nrow = 16L, ncol = 16L)
is_bm_bitmap(space_matrix)
space_glyph <- bm_bitmap(space_matrix)
is_bm_bitmap(space_glyph)
}
\seealso{
\code{\link[=bm_bitmap]{bm_bitmap()}}
}
