% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_font.R
\name{is_bm_font}
\alias{is_bm_font}
\title{Test if the object is a bitmap font object}
\usage{
is_bm_font(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{is_bm_font()} returns \code{TRUE} for \code{bm_font} objects (or subclasses)
and \code{FALSE} for all other objects.
}
\examples{
 font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
 font <- read_hex(font_file)
 is_bm_font(font)
}
\seealso{
\code{\link[=bm_font]{bm_font()}}
}
