% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_glyph.R
\name{is_bm_glyph}
\alias{is_bm_glyph}
\title{Test if the object is a bitmap glyph object}
\usage{
is_bm_glyph(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{is_bm_glyph()} returns \code{TRUE} for \code{bm_glyph} objects (or subclasses)
and \code{FALSE} for all other objects.
}
\examples{
 space_matrix <- matrix(0, nrow = 16, ncol = 16)
 space_glyph <- bm_glyph(space_matrix)
 is_bm_glyph(space_glyph)
 is_bm_glyph(space_matrix)
}
\seealso{
\code{\link[=bm_glyph]{bm_glyph()}}
}
