% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bm_matrix.R
\name{is_bm_pixmap}
\alias{is_bm_pixmap}
\title{Test if the object is a pixmap object}
\usage{
is_bm_pixmap(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{is_bm_pixmap()} returns \code{TRUE} for \code{bm_pixmap} objects (or subclasses)
and \code{FALSE} for all other objects.
}
\examples{
pm <- bm_pixmap(matrix(c("red", "blue", "green", "black"),
                nrow = 2L, byrow = TRUE))
is_bm_pixmap(pm)
}
\seealso{
\code{\link[=bm_pixmap]{bm_pixmap()}}, \code{\link[=as_bm_pixmap]{as_bm_pixmap()}}
}
