% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths.R
\name{bm_widths}
\alias{bm_widths}
\alias{bm_heights}
\title{Widths or heights of bitmaps in bitmap lists}
\usage{
bm_widths(x, unique = TRUE)

bm_heights(x, unique = TRUE)
}
\arguments{
\item{x}{A \code{bm_list()} object including a \code{bm_font()} object.}

\item{unique}{Apply \code{\link[base:unique]{base::unique()}} to the returned integer vector.}
}
\value{
A integer vector of the relevant length of each
of the \code{bm_bitmap()} objects in \code{x}.
If \code{unique} is \code{TRUE} then any duplicates will have been removed.
}
\description{
\code{bm_widths()} returns the widths of the bitmaps in a \code{bm_list()} while
\code{bm_heights()} returns the heights of the bitmaps.
}
\examples{
  font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
  font <- read_hex(font_file)
  bm_widths(font) # every glyph in the font is 8 pixels wide
  bm_heights(font) # every glyph in the font is 16 pixels high
}
