% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bm_bitmap.R
\name{plot.bm_bitmap}
\alias{plot.bm_bitmap}
\title{Plot bitmap object}
\usage{
\method{plot}{bm_bitmap}(x, ..., col = c("grey80", "black", "grey40"), interpolate = FALSE)
}
\arguments{
\item{x}{A \code{bm_bitmap()} object}

\item{...}{Passed to \code{\link[grid:grid.raster]{grid::grid.raster()}}.}

\item{col}{Character vector of R color specifications. Passed to \code{\link[=as.raster.bm_bitmap]{as.raster.bm_bitmap()}}.}

\item{interpolate}{Passed to \code{\link[grid:grid.raster]{grid::grid.raster()}}.}
}
\value{
A \code{grid} rastergrob grob object silently.
As a side effect will draw to graphics device.
}
\description{
\code{plot.bm_bitmap()} plots a bitmap object to the graphics device.
}
\examples{
  font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
  font <- read_hex(font_file)
  capital_r <- font[[str2ucp("R")]]

  plot(capital_r)

  grid::grid.newpage()
  plot(capital_r, col = c("yellow", "blue"))
}
\seealso{
\code{\link[=bm_bitmap]{bm_bitmap()}}, \code{\link[=as.raster.bm_bitmap]{as.raster.bm_bitmap()}}
}
