% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bm_glyph.R
\name{plot.bm_glyph}
\alias{plot.bm_glyph}
\title{Plot bitmap glyph object}
\usage{
\method{plot}{bm_glyph}(x, ..., col = c("grey80", "black", "grey40"), interpolate = FALSE)
}
\arguments{
\item{x}{A \code{bm_glyph()} object}

\item{...}{Passed to \code{\link[grid:grid.raster]{grid::grid.raster()}}.}

\item{col}{Character vector of R color specifications. Passed to \code{\link[=as.raster.bm_glyph]{as.raster.bm_glyph()}}.}

\item{interpolate}{Passed to \code{\link[grid:grid.raster]{grid::grid.raster()}}.}
}
\value{
A \code{grid} rastergrob grob object silently.
As a side effect will draw to graphics device.
}
\description{
\code{plot.bm_glyph()} plots a bitmap glyph object to the graphics device.
}
\examples{
  font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
  font <- read_hex(font_file)
  capital_r <- font[[code_point_from_char("R")]]

  plot(capital_r)

  grid::grid.newpage()
  plot(capital_r, col = c("yellow", "blue"))
}
\seealso{
\code{\link[=bm_glyph]{bm_glyph()}}, \code{\link[=as.raster.bm_glyph]{as.raster.bm_glyph()}}
}
