% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bm_matrix.R
\name{plot.bm_matrix}
\alias{plot.bm_matrix}
\alias{plot.bm_bitmap}
\alias{plot.bm_pixmap}
\alias{as.raster.bm_bitmap}
\alias{as.raster.bm_pixmap}
\title{Plot bitmap/pixmap objects}
\usage{
\method{plot}{bm_bitmap}(x, ..., col = c("grey80", "black", "grey40"), interpolate = FALSE)

\method{plot}{bm_pixmap}(x, ..., interpolate = FALSE)

\method{as.raster}{bm_bitmap}(x, ..., col = c("grey80", "black", "grey40"))

\method{as.raster}{bm_pixmap}(x, ...)
}
\arguments{
\item{x}{A \code{bm_bitmap()} object}

\item{...}{Passed to \code{\link[grid:grid.raster]{grid::grid.raster()}}.}

\item{col}{Character vector of R color specifications.
First color is used for values equal to 0, second color for values equal to 1, etc.}

\item{interpolate}{Passed to \code{\link[grid:grid.raster]{grid::grid.raster()}}.}
}
\value{
\code{plot.bm_bitmap()} and \code{plot.bm_pixmap()} return a \code{\link[grid:grid.raster]{grid::rasterGrob()}} object silently.
As a side effect will draw to graphics device.
\code{as.raster.bm_bitmap()} and \code{as.raster.bm_pixmap()} return "raster" objects (see \code{\link[grDevices:as.raster]{grDevices::as.raster()}}).
}
\description{
\code{plot.bm_bitmap()} plots a \code{\link[=bm_bitmap]{bm_bitmap()}} object to the graphics device
while \code{plot.bm_pixmap()} plots a \code{\link[=bm_pixmap]{bm_pixmap()}} object to the graphics device.
They are wrappers around \code{\link[grid:grid.raster]{grid::grid.raster()}} and \code{as.raster.bm_bitmap()}
or \code{as.raster.bm_pixmap()}.
which converts a bitmap glyph object to a raster object.
}
\examples{
  font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
  font <- read_hex(font_file)
  capital_r <- bm_extend(font[[str2ucp("R")]], left = 1L)
  capital_r <- bm_extend(capital_r, sides = 1L, value = 2L)  # add a border effect

  plot(capital_r)

  grid::grid.newpage()
  plot(capital_r, col = c("yellow", "blue", "red"))
}
\seealso{
\code{\link[=bm_bitmap]{bm_bitmap()}}, \code{\link[=bm_pixmap]{bm_pixmap()}}
}
