% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bm_bitmap.R
\name{print.bm_bitmap}
\alias{print.bm_bitmap}
\title{Print bitmap objects}
\usage{
\method{print}{bm_bitmap}(x, ..., labels = c("\u2591", "\u2588", "\u2592"))
}
\arguments{
\item{x}{A \code{bm_bitmap()} object}

\item{...}{Further arguments passed to or from other methods.}

\item{labels}{Character vector of the labels to use for each integer value i.e.
The first character for integer \code{0L},
the second character for integer \code{1L}, and so on.}
}
\value{
A character vector of the string representation invisibly.
}
\description{
\code{print.bm_bitmap()} prints a representation of bitmap objects to the terminal
}
\examples{
  plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
  plus_sign[5L, 3:7] <- 1L
  plus_sign[3:7, 5L] <- 1L
  plus_sign <- bm_bitmap(plus_sign)
  print(plus_sign, labels = c(" ", "#"))

  space_glyph <- bm_bitmap(matrix(0L, nrow = 8L, ncol = 8L))
  print(space_glyph, labels = ".")
}
\seealso{
\code{\link[=bm_bitmap]{bm_bitmap()}}
}
