% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hex.R
\name{read_hex}
\alias{read_hex}
\title{Read hex format bitmap font files}
\usage{
read_hex(con)
}
\arguments{
\item{con}{A connection object or a character string.
This can be a compressed file.
See \code{\link[base:readLines]{base::readLines()}} for more info.}
}
\description{
\code{read_hex()} reads in hex format bitmap font files
as a \code{\link[=bm_font]{bm_font()}} object.
}
\examples{
 font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
 font <- read_hex(font_file)
 capital_r <- font[[code_point_from_name("LATIN CAPITAL LETTER R")]]
 print(capital_r, labels = c(".", "#"))
}
\seealso{
\code{\link[=bm_font]{bm_font()}}
}
