% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-unicode.R
\name{ucp2label}
\alias{ucp2label}
\alias{is_combining_character}
\title{Other Unicode utilities}
\usage{
ucp2label(x)

is_combining_character(x)
}
\arguments{
\item{x}{A character vector of Unicode code points.}
}
\value{
\code{ucp2name()} returns a character vector of Unicode labels.
}
\description{
\code{ucp2label()} returns Unicode code point \dQuote{labels} as a character vector.
\code{is_combining_character()} returns \code{TRUE} if the character is a \dQuote{combining} character.
}
\examples{
  # Get the Unicode Code Point "label" for "R"
  ucp2label(str2ucp("R"))

  is_combining_character(str2ucp("a"))
  is_combining_character(str2ucp("\u0300")) # COMBINING GRAVE ACCENT
}
\seealso{
\code{\link[=hex2ucp]{hex2ucp()}}, \code{\link[=int2ucp]{int2ucp()}}, \code{\link[=name2ucp]{name2ucp()}}, and \code{\link[=str2ucp]{str2ucp()}} all return Unicode code points.
}
