% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.raster.bm_bitmap.R
\name{as.raster.bm_bitmap}
\alias{as.raster.bm_bitmap}
\title{Coerce bitmap glyph object to raster object}
\usage{
\method{as.raster}{bm_bitmap}(x, ..., col = c("grey80", "black", "grey40"))
}
\arguments{
\item{x}{A \code{bm_bitmap()} object}

\item{...}{Further arguments passed to or from other methods.}

\item{col}{Character vector of R color specifications}
}
\description{
\code{as.raster.bm_bitmap()} converts a bitmap glyph object to a raster object.
}
\examples{
  font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
  font <- read_hex(font_file)
  capital_r <- font[[str2ucp("R")]]
  grDevices::as.raster(capital_r)
}
