% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_list.R
\name{as_bm_list}
\alias{as_bm_list}
\alias{as_bm_list.default}
\alias{as_bm_list.list}
\title{Coerce to bitmap glyph list objects}
\usage{
as_bm_list(x, ...)

\method{as_bm_list}{default}(x, ...)

\method{as_bm_list}{list}(x, ...)
}
\arguments{
\item{x}{An object that can reasonably be coerced to a \code{\link[=bm_list]{bm_list()}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{\link[=bm_list]{bm_list()}} object.
}
\description{
\code{as_bm_list()} turns an existing object into a \code{\link[=bm_list]{bm_list()}} object.
}
\examples{
  plus_sign <- matrix(0L, nrow = 9L, ncol = 9L)
  plus_sign[5L, 3:7] <- 1L
  plus_sign[3:7, 5L] <- 1L
  plus_sign_glyph <- bm_glyph(plus_sign)
  plus_sign_code_point <- code_point_from_name("PLUS SIGN") # "U+002B"

  space_glyph <- bm_glyph(matrix(0L, nrow = 9L, ncol = 9L))
  space_code_point <- code_point_from_name("SPACE") # "U+0020"

  l <- list()
  l[[plus_sign_code_point]] <- plus_sign_glyph
  l[[space_code_point]] <- space_glyph
  bl <- as_bm_list(l)
  is_bm_list(bl)

}
\seealso{
\code{\link[=bm_list]{bm_list()}}
}
