% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_clamp.R
\name{bm_clamp}
\alias{bm_clamp}
\title{Clamp bitmap glyph integer values.}
\usage{
bm_clamp(bm_object, lower = 0L, upper = 1L)
}
\arguments{
\item{bm_object}{Either a \code{\link[=bm_glyph]{bm_glyph()}} or \code{\link[=bm_list]{bm_list()}} object (including \code{\link[=bm_font]{bm_font()}}).}

\item{lower}{Integer value.  Any value below \code{lower} will be clamped to \code{lower}.}

\item{upper}{Integer value.  Any value above \code{upper} will be clamped to \code{upper}.}
}
\description{
\code{bm_clamp()} \dQuote{clamps} bitmap glyph integers to lie between a lower and upper value.
The default lower and upper values coerce the bitmap glyph
into a binary bitmap glyph (as expected by most bitmap font formats).
}
