% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_point.R
\name{code_point}
\alias{code_point}
\alias{code_point_from_name}
\title{Get Unicode code points}
\usage{
code_point(x)

code_point_from_name(x)
}
\arguments{
\item{x}{R objects coercible to the respective Unicode character data types.
See \code{\link[Unicode:u_char_basics]{Unicode::as.u_char()}} for \code{code_point()} and
\code{\link[Unicode:u_char_names]{Unicode::u_char_from_name()}} for \code{code_point_from_name()}.}
}
\value{
A character vector of Unicode code points.
}
\description{
\code{code_point()} and \code{code_point_from_name()} return Unicode code points.
}
\details{
\code{code_point()} is an alias for \code{as.character(Unicode::as.u_char(x))}.
\code{code_point_from_name(x)} is an alias for \code{as.character(Unicode::u_char_from_name(x))}.
The names of \code{bm_font()} objects must be character vectors and not \code{u_char} objects.
}
\examples{
  code_point("20")
  code_point("U+0020")
  code_point("0x0020")
  code_point_from_name("SPACE")
  code_point_from_name("LATIN CAPITAL LETTER R")
}
