% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_list.R
\name{is_bm_list}
\alias{is_bm_list}
\title{Test if the object is a bitmap glyph list object}
\usage{
is_bm_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{is_bm_list()} returns \code{TRUE} for \code{\link[=bm_list]{bm_list()}} objects (or subclasses)
and \code{FALSE} for all other objects.
}
\examples{
 font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
 font <- read_hex(font_file)
 is_bm_font(font)
}
\seealso{
\code{\link[=bm_list]{bm_list()}}
}
